+++ src/Flare.Theme.FluentUI2.Tokens/FluentUI2Tokens.cs
@@ -1189,6 +1189,7 @@ public class FluentUI2Tokens
     internal static readonly PopoverTokens Popover = new()
     {
         Radius = "var(--flare-shape-small)",
+        Offset = "0.5rem",
     };
 
     internal static readonly AvatarTokens Avatar = new()