+++ src/Flare.Theme.MaterialDesign3.Tokens/MaterialDesign3Tokens.cs
@@ -926,6 +926,7 @@ public class MaterialDesign3Tokens
     internal static readonly PopoverTokens Popover = new()
     {
         Radius = "var(--flare-shape-medium)",
+        Offset = "0.5rem",
     };
 
     internal static readonly AvatarTokens Avatar = new()